requirements.txt
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
django-sendgrid-webhook==0.3.0
idna==2.10
metabasepy==1.9.0
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
starkbank-ecdsa==1.1.0
urllib3==1.26.4
webdriver-manager==3.4.1
